feat: 설정 로그아웃 추가
