+++ src/hooks/setting/settingHook.ts
@@ -1,8 +1,10 @@
 import React, { useState } from 'react';
+import useLogout from '@hooks/login/logoutHook';
 import { Alert, Linking } from 'react-native';
 
 const useSetting = () => {
   const [deleteModalVisible, setDeleteModalVisible] = useState(false);
+  const { handleLogout } = useLogout();
 
   const onFeedback = async () => {
     const url = 'https://forms.gle/nUCrj6JLNCnU8Dez6'; // env에 추가하기
@@ -18,16 +20,12 @@ const useSetting = () => {
     setDeleteModalVisible(true);
   };
 
-  const onLogout = () => {
-    console.log('logout');
-  };
-
   return {
     deleteModalVisible,
     setDeleteModalVisible,
     onFeedback,
     onDeleteAccount,
-    onLogout,
+    handleLogout,
   };
 };
